Move the taylor series initialisation inside Operators
1. Rename examples/operators.py -> examples/fwi_operators.py
2. The taylor expansion code and hence the generated stencil can now be
parametrised for 2D/3D as well as higher order stencils.
3. AcousticWave now has nearly no code in it.
